+++ src/main/java/snw/rfm/api/events/HunterCatchPlayerEvent.java
@@ -77,6 +77,8 @@ public final class HunterCatchPlayerEvent extends Event implements Cancellable {
 
     /**
      * 设置该玩家最终能获得的 B币 。
+     * <p>
+     * <b>注意，此方法被成功调用后，无论向 {@link HunterCatchPlayerEvent#getCoinEarned(boolean)} 方法传递 <code>true</code> 还是 <code>false</code> ，返回值都始终会是提供给此方法的值。</b>
      *
      * @param coin 玩家将获得的 B币 。
      * @throws IllegalArgumentException 当 <code>coin</code> 是负数或是 NaN 时抛出